refactor: change users route name to auth route
