+++ src/routes/index.ts
@@ -1,12 +1,12 @@
 import { Router } from "express";
-import usersRouter from "./usersRouter";
+import authRouter from "./authRouter";
 import validateToken from "../middlewares/tokenValidatorMiddleware";
 import credentialsRouter from "./credentialsRouter";
 // import paymentsRouter from "./paymentsRouter";
 
 const router: Router = Router();
 
-router.use(usersRouter);
+router.use(authRouter);
 router.use(validateToken);
 router.use(credentialsRouter);
 // router.use(paymentsRouter);
